+++ common/src/main/java/de/cech12/usefulhats/Constants.java
@@ -20,7 +20,6 @@ public class Constants {
     /** Logger instance */
     public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
 
-    public static final String CURIOS_MOD_ID = "curios";
     public static final String LUCENT_MOD_ID = "lucent";
 
     public static Supplier<DataComponentType<EnderHelmetItem.Position>> ENDER_HELMET_POSITION;